requirements.txt
diff-match-patch==20121119
Django==1.8.2
django-import-export==0.2.7
tablib==0.10.0
